fix: service register headset broadcastreceiver
Fix #43
