+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,6 @@
     <uses-permission android:name="android.permission.BLUETOOTH_ADMIN" />
 
     <application
-        android:name=".MusicPlayerApplication"
         android:allowBackup="true"
         android:icon="@drawable/player_icon"
         android:label="@string/app_name"
